Publishing NewCustomer-Event
Now, when creating a new customer, also publishs an event to
"http://customer-event-log/events". 
"customer-event-log" is the *name* another microservices is registered
within the eureka service registry.
Now publishing to "http://customer-event-log/events" via the
RestTemplate.
Good to know that (because of spring cloud) this is a special (load
balanced) RestTemplate that uses the DiscoveryClient to obtain the real
hostname and port and then makes the real request (to
http://localhost:8095/events for example).